src/requirements.txt
cloudshell-shell-core>=5.0.4,<6.0
cloudshell-cp-core>=1.0.4.89,<1.1
cloudshell-automation-api
requests
